+++ res.lua
@@ -1,0 +1,19 @@
+return {
+  absolute_offset = 66,
+  line_number = 10,
+  lines = {
+    text = "[[1746537858-CRIT|test2]]\n",
+  },
+  path = {
+    text = "/home/n451/Plugins/obsidian.nvim/.repro/vault/1746537981-USPT.md",
+  },
+  submatches = {
+    {
+      ["end"] = 18,
+      match = {
+        text = "[[1746537858-CRIT|",
+      },
+      start = 0,
+    },
+  },
+}