+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove `itertools.lua` in favor of `vim.iter`
 - Configure `foldmethod`, `foldexpr`, and `foldlevel` in ftplugin instead of the BufEnter event. This allows user overrides of these configurations.
 - Remove `fillchars`, `foldtext`, and `smoothscroll` configurations.
+- Relaxed yaml dump quoting condition so that dates are not quoted
 
 ### Fixed